Clarify backend package ownership (#32)
* Document backend package organization candidate

* Clarify backend package ownership

* Document plugin adapter consolidation

* Consolidate plugin adapters

* Remove review planning documents

* Fix HTTP API package documentation

* Clarify state and migration ownership

* Internalize rules and observability

* Simplify backend import aliases

* Decompose server application wiring

* Make migrations startup-owned

* Reduce backend lint suppressions

* Use native PostgreSQL pool size type

* Preserve PostgreSQL row count types

+++ backend/internal/store/models.go
@@ -174,7 +174,7 @@ type MerchantCategoryEntry struct {
 type SyncStatus struct {
 	LastSyncedAt   *time.Time `json:"last_synced_at"`
 	Error          *string    `json:"error"`
-	EntriesUpdated int        `json:"entries_updated"`
+	EntriesUpdated int64      `json:"entries_updated"`
 }
 
 // RuleRow is a rule as stored in the database.
